fix(app): guard non-integer volume numbers in stats volume gaps
Only treat whole-number volume numbers as gap candidates in the generic stats
dashboard so fractional volumeNumber values no longer produce bogus missing-volume
alerts. Update the test to construct volumeNumber as a double to match the model.

+++ test/features/library/stats/stats_dashboard_test.dart
@@ -229,7 +229,7 @@ void main() {
               title: 'Vinland Saga',
               series: CatalogSeriesDetails(
                 seriesTitle: 'Vinland Saga',
-                volumeNumber: volume,
+                volumeNumber: volume.toDouble(),
               ),
             ),
             ownedItem: OwnedItem(
